Use base class for <FileType>Exists type methods
The only thing that changes is the end of the url
